feat: Implementar outras páginas e demias correções de Bugs.

+++ models/Alunos.py
@@ -1,11 +1,11 @@
 class Alunos:
-    def __init__(self, codigo=0, nome=0, codCidade=0, nascimento=0, peso=0, altura=0):
+    def __init__(self, codigo=0, nome=0, codCidade=0, nascimento=0, peso=0, altura=0.0):
         self.codigo = int(codigo)
         self.nome = str(nome)
         self.codCidade = int(codCidade)
         self.nascimento = str(nascimento)
         self.peso = int(peso)
-        self.altura = int(altura)
+        self.altura = float(altura)
 
     def __str__(self):
         return f"{self.codigo};{self.nome};{self.codCidade};{self.nascimento};{self.peso};{self.altura}"
